pyproject.toml
[tool.poetry]
name = "rl_test"
version = "0.1.0"
description = "implements reinforcement learning"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
gym = "^0.17.2"
numpy-quaternion = "^2020.5.19"
matplotlib = "^3.2.2"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
